+++ src/config.py
@@ -15,8 +15,8 @@ class User:
     PASSWORD: "123456",
   }
   GLANCE = {
-    ACCOUNT: "keystone"
-    PASSWORD: "123456"
+    ACCOUNT: "keystone",
+    PASSWORD: "123456",
   }
 
 class Agent:
@@ -33,25 +33,4 @@ class Hosts:
     Agent.COMPUTE: "127.0.0.1",
   }
 
-class Task:
-  def __init__(self, info, command):
-    self.info = info
-    self.command  = command
-  def info(self):
-    return self.info
-  def command(self):
-    return self.command
-  def exe(self):
-    print "[INFO] " + self.info
-    subprocess.call(self.command.split())
 
-class Tee(object):
-    def __init__(self, *files):
-        self.files = files
-    def write(self, obj):
-        for f in self.files:
-            f.write(obj)
-            f.flush() # If you want the output to be visible immediately
-    def flush(self) :
-        for f in self.files:
-            f.flush()
